daemon/server/backend: move build options to buildbackend
